DBZ-2552 Refactored database state capture logic
