+++ debezium-connector-oracle/src/main/java/io/debezium/connector/oracle/logminer/LogMinerDatabaseStateWriter.java
@@ -1,0 +1,99 @@
+/*
+ * Copyright Debezium Authors.
+ *
+ * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
+ */
+package io.debezium.connector.oracle.logminer;
+
+import java.sql.SQLException;
+import java.util.ArrayList;
+import java.util.List;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import io.debezium.connector.oracle.OracleConnection;
+
+/**
+ * Utility class that can write the contents of several Oracle LogMiner tables to the
+ * connector's log if and only if DEBUG logging is enabled.
+ *
+ * @author Chris Cranford
+ */
+public class LogMinerDatabaseStateWriter {
+
+    private static final Logger LOGGER = LoggerFactory.getLogger(LogMinerDatabaseStateWriter.class);
+
+    public static void write(OracleConnection connection) {
+        if (LOGGER.isDebugEnabled()) {
+            LOGGER.debug("Configured redo logs are:");
+            try {
+                logQueryResults(connection, "SELECT * FROM V$LOGFILE");
+            }
+            catch (SQLException e) {
+                LOGGER.debug("Failed to obtain redo log table entries", e);
+            }
+            LOGGER.debug("Archive logs for last 48 hours:");
+            try {
+                logQueryResults(connection, "SELECT * FROM V$ARCHIVED_LOG WHERE FIST_TIME >= SYSDATE - 2");
+            }
+            catch (SQLException e) {
+                LOGGER.debug("Failed to obtain archive log table entries", e);
+            }
+            LOGGER.debug("Available logs are:");
+            try {
+                logQueryResults(connection, "SELECT * FROM V$LOG");
+            }
+            catch (SQLException e) {
+                LOGGER.debug("Failed to obtain log table entries", e);
+            }
+            LOGGER.debug("Log history last 24 hours:");
+            try {
+                logQueryResults(connection, "SELECT * FROM V$LOG_HISTORY WHERE FIRST_TIME >= SYSDATE - 1");
+            }
+            catch (SQLException e) {
+                LOGGER.debug("Failed to obtain log history", e);
+            }
+            LOGGER.debug("Log entries registered with LogMiner are:");
+            try {
+                logQueryResults(connection, "SELECT * FROM V$LOGMNR_LOGS");
+            }
+            catch (SQLException e) {
+                LOGGER.debug("Failed to obtain registered logs with LogMiner", e);
+            }
+            LOGGER.debug("Log mining session parameters are:");
+            try {
+                logQueryResults(connection, "SELECT * FROM V$LOGMNR_PARAMETERS");
+            }
+            catch (SQLException e) {
+                LOGGER.debug("Failed to obtain log mining session parameters", e);
+            }
+        }
+    }
+
+    /**
+     * Helper method that dumps the result set of an arbitrary SQL query to the connector's logs.
+     *
+     * @param connection the database connection
+     * @param query the query to execute
+     * @throws SQLException thrown if an exception occurs performing a SQL operation
+     */
+    private static void logQueryResults(OracleConnection connection, String query) throws SQLException {
+        connection.query(query, rs -> {
+            int columns = rs.getMetaData().getColumnCount();
+            List<String> columnNames = new ArrayList<>();
+            for (int index = 1; index <= columns; ++index) {
+                columnNames.add(rs.getMetaData().getColumnName(index));
+            }
+            LOGGER.debug("{}", columnNames);
+            while (rs.next()) {
+                List<Object> columnValues = new ArrayList<>();
+                for (int index = 1; index <= columns; ++index) {
+                    columnValues.add(rs.getObject(index));
+                }
+                LOGGER.debug("{}", columnValues);
+            }
+        });
+    }
+
+}
