+++ debezium-connector-oracle/src/main/java/io/debezium/connector/oracle/logminer/LogMinerHelper.java
@@ -162,84 +162,6 @@ public class LogMinerHelper {
     }
 
     /**
-     * Helper method that will dump the state of various critical tables used by the LogMiner implementation
-     * to derive state about which logs are to be mined and processed by the Oracle LogMiner session.
-     *
-     * @param connection the database connection
-     */
-    public static void logDatabaseState(OracleConnection connection) {
-        if (LOGGER.isDebugEnabled()) {
-            LOGGER.debug("Configured redo logs are:");
-            try {
-                logQueryResults(connection, "SELECT * FROM V$LOGFILE");
-            }
-            catch (SQLException e) {
-                LOGGER.debug("Failed to obtain redo log table entries", e);
-            }
-            LOGGER.debug("Available archive logs are:");
-            try {
-                logQueryResults(connection, "SELECT * FROM V$ARCHIVED_LOG");
-            }
-            catch (SQLException e) {
-                LOGGER.debug("Failed to obtain archive log table entries", e);
-            }
-            LOGGER.debug("Available logs are:");
-            try {
-                logQueryResults(connection, "SELECT * FROM V$LOG");
-            }
-            catch (SQLException e) {
-                LOGGER.debug("Failed to obtain log table entries", e);
-            }
-            LOGGER.debug("Log history last 24 hours:");
-            try {
-                logQueryResults(connection, "SELECT * FROM V$LOG_HISTORY WHERE FIRST_TIME >= SYSDATE - 1");
-            }
-            catch (SQLException e) {
-                LOGGER.debug("Failed to obtain log history", e);
-            }
-            LOGGER.debug("Log entries registered with LogMiner are:");
-            try {
-                logQueryResults(connection, "SELECT * FROM V$LOGMNR_LOGS");
-            }
-            catch (SQLException e) {
-                LOGGER.debug("Failed to obtain registered logs with LogMiner", e);
-            }
-            LOGGER.debug("Log mining session parameters are:");
-            try {
-                logQueryResults(connection, "SELECT * FROM V$LOGMNR_PARAMETERS");
-            }
-            catch (SQLException e) {
-                LOGGER.debug("Failed to obtain log mining session parameters", e);
-            }
-        }
-    }
-
-    /**
-     * Helper method that dumps the result set of an arbitrary SQL query to the connector's logs.
-     *
-     * @param connection the database connection
-     * @param query the query to execute
-     * @throws SQLException thrown if an exception occurs performing a SQL operation
-     */
-    private static void logQueryResults(OracleConnection connection, String query) throws SQLException {
-        connection.query(query, rs -> {
-            int columns = rs.getMetaData().getColumnCount();
-            List<String> columnNames = new ArrayList<>();
-            for (int index = 1; index <= columns; ++index) {
-                columnNames.add(rs.getMetaData().getColumnName(index));
-            }
-            LOGGER.debug("{}", columnNames);
-            while (rs.next()) {
-                List<Object> columnValues = new ArrayList<>();
-                for (int index = 1; index <= columns; ++index) {
-                    columnValues.add(rs.getObject(index));
-                }
-                LOGGER.debug("{}", columnValues);
-            }
-        });
-    }
-
-    /**
      * This method removes all added log files from mining
      * @param conn connection
      * @throws SQLException something happened
